dmaengine: qcom: bam_dma: Add remote power collapse mode
In some configurations, the BAM DMA controller is set up by a remote
processor and the local processor can simply start making use of it
without setting up the BAM. This is already supported using the
"qcom,controlled-remotely" property.

However, for some reason another possible configuration is that the
remote processor is responsible for powering up the BAM, but we are
still responsible for initializing it (e.g. resetting it etc).

This configuration is quite challenging to handle properly because
the power control is handled through separate channels
(e.g. device-specific SMSM interrupts / smem-states). Great care
must be taken to ensure the BAM registers are not accessed while
the BAM is power-collapsed since this results in a bus stall.

Attempt to support this configuration with minimal device-specific
code in the bam_dma driver by tracking the number of requested
channels. Consumers of DMA channels are responsible to only request
DMA channels when the BAM was powered on by the remote processor,
and to release them before the BAM is power-collapsed.

When the first channel is requested the BAM is initialized (reset)
and it is also put into reset when the last channel was released.

Signed-off-by: Stephan Gerhold <stephan@gerhold.net>
---
NOTE: This is *not* a compile-time requirement for the BAM-DMUX driver
      so this could also go through the dmaengine tree.

I tried to come up with other solutions for this situation, but this
is the cleanest I came up with so far. The main advantage is that
it keeps the bam_dma driver generic and fairly simple. The main
disadvantage might be that there is some overhead when the DMA channels
are repeatedly requested and released (the BAM-DMUX driver uses
PM runtime to autosuspend after 1 second of inactivity).

Some alternative ideas (but I'm not sure how they would work exactly):

  - Have some dmaengine_*() operation to make the bam_dma driver aware
    that the BAM is power-collapsed (instead of requesting/releasing
    the channels every time).

  - Give the BAM-DMUX power control IRQ to bam_dma so it knows when
    the BAM is power-collapsed or not.

    - Perhaps even give the smem-state to bam_dma so it could request
      powering on the BAM itself. This would be quite ugly though,
      since bam_dma already uses runtime PM but differently (only active
      during register writes, not necessarily active during transfers).

    Note however that the power control of BAM-DMUX also involves
    queuing RX buffers so there would be much more coordination
    needed between bam_dma and bam-dmux.