arm64: dts: qcom: msm8916-longcheer-l8150: Allow flash led brightness control
L8150 uses SGM3785 Flash LED driver. It is similar to SGM3140 but
can also be controlled with PWM. It's possible to get PWM signal on
the enable pin from GP_CLK1 clock that can be muxed to pin 32.
Use this clock with the new pwm-clk driver to enable brightness control.

Signed-off-by: Nikita Travkin <nikita@trvn.ru>